# Use node-glob's `ignore` option, massive perf improvement regarding negative globs