--- a/index.js
+++ b/index.js
@@ -31,34 +31,22 @@ module.exports = function (patterns, opts, cb) {
 		return;
 	}
 
-	negatives.forEach(function (negative) {
-		negative.matcher = new Minimatch(negative.pattern, opts);
-	});
-
 	async.parallel(positives.map(function (positive) {
 		return function (cb2) {
+			var negativePatterns = negatives.filter(function (negative) {
+				return negative.index > positive.index;
+			}).map(function (negative) {
+				return negative.pattern.slice(1);
+			});
+
+			opts.ignore = (opts.ignore || []).concat(negativePatterns);
+
 			glob(positive.pattern, opts, function (err, paths) {
 				if (err) {
 					cb2(err);
 					return;
 				}
-
-				var negativeMatchers = negatives.filter(function (negative) {
-					return negative.index > positive.index;
-				}).map(function (negative) {
-					return negative.matcher;
-				});
-
-				if (negativeMatchers.length === 0) {
-					cb2(null, paths);
-					return;
-				}
-
-				cb2(null, paths.filter(function (path) {
-					return negativeMatchers.every(function (matcher) {
-						return matcher.match(path);
-					});
-				}));
+				cb2(null, paths);
 			});
 		};
 	}), function (err, paths) {