--- a/package.json
+++ b/package.json
@@ -53,7 +53,7 @@
   "dependencies": {
     "array-union": "^1.0.1",
     "async": "^0.9.0",
-    "glob": "^4.0.2",
+    "glob": "vegetableman/node-glob#fix-ignore-branch",
     "minimatch": "^2.0.1"
   },
   "devDependencies": {