--- a/package.json
+++ b/package.json
@@ -53,8 +53,7 @@
   "dependencies": {
     "array-union": "^1.0.1",
     "async": "^0.9.0",
-    "glob": "vegetableman/node-glob#fix-ignore-branch",
-    "minimatch": "^2.0.1"
+    "glob": "vegetableman/node-glob#fix-ignore-branch"
   },
   "devDependencies": {
     "glob-stream": "wearefractal/glob-stream#master",