--- a/package.json
+++ b/package.json
@@ -53,7 +53,8 @@
   "dependencies": {
     "array-union": "^1.0.1",
     "async": "^0.9.0",
-    "glob": "vegetableman/node-glob#fix-ignore-branch"
+    "glob": "vegetableman/node-glob#fix-ignore-branch",
+    "object-assign": "^2.0.0"
   },
   "devDependencies": {
     "glob-stream": "wearefractal/glob-stream#master",