--- a/test.js
+++ b/test.js
@@ -58,3 +58,22 @@ it('cwd option', function () {
 	assert.deepEqual(globby.sync(['a.tmp', '*.tmp', '!{c,d,e}.tmp'], {cwd: cwd}), ['a.tmp', 'b.tmp']);
 	process.chdir(cwd);
 });
+
+it('should not mutate the options object - async', function(cb) {
+	var ignore = [];
+	var opts = { ignore: ignore };
+	globby(['*.tmp', '!b.tmp'], opts, function (err, paths) {
+		assert(!err, err);
+		assert.strictEqual(opts.ignore, ignore); // same reference
+		assert.deepEqual(opts.ignore, ignore); // same contents
+		cb();
+	});
+});
+
+it('should not mutate the options object - sync', function() {
+	var ignore = [];
+	var opts = { ignore: ignore };
+	globby.sync(['*.tmp', '!b.tmp'], opts);
+	assert.strictEqual(opts.ignore, ignore); // same reference
+	assert.deepEqual(opts.ignore, ignore); // same contents
+});