Benchmark results on my main dev machine:

```
                      negative globs (whole dir)
          10,296 op/s » globby async (working directory)
             100 op/s » globby async (upstream/master)
          20,547 op/s » globby sync (working directory)
             105 op/s » globby sync (upstream/master)
              66 op/s » glob-stream
```

@sindresorhus PTAL.