@sindresorhus 2015-02-23
Oh wow. That's an awesome improvement!