@sindresorhus 2015-02-23
![tumblr_mqmk6j8kkz1qdlh1io1_250](https://cloud.githubusercontent.com/assets/170270/6323148/4dc4877c-bb57-11e4-924d-0b672cfa7a89.gif)